dom.js, storeMenu.js 파일분리, initEventListeners 함수통합
