+++ src/js/utils/dom.js
@@ -1,0 +1,3 @@
+// JS에서 DOM element 가져올 때 관용적으로 $표시를 사용한다.
+// $표시로 DOM element return해서 반복 줄이는 함수
+export const $ = (selector) => document.querySelector(selector);
